Move increment over to the utils folder
